Create separate storage class
