+++ src/util/block-site.js
@@ -2,42 +2,7 @@
 import Autolinker from 'autolinker';
 import * as UrlParser from 'url-parse';
 import { message } from 'antd';
-
-let listeners = [];
-
-function getFromStorage(...keys) {
-    return new Promise(resolve => {
-        if (chrome && chrome.storage) {
-            chrome.storage.sync.get(keys, result => {
-                resolve(result);
-            });
-        } else {
-            let result = keys.reduce((acc, val) => {
-                acc[val] = JSON.parse(localStorage.getItem(val));
-                return acc;
-            }, {});
-
-            resolve(result);
-        }
-    });
-}
-
-function setInStorage(items) {
-    return new Promise(resolve => {
-        if (chrome && chrome.storage) {
-            chrome.storage.sync.set(items, () => {
-                resolve();
-            });
-        } else {
-            Object.keys(items).forEach(key => {
-                localStorage.setItem(key, JSON.stringify(items[key]));
-            });
-            listeners.forEach(callback => callback());
-
-            resolve();
-        }
-});
-}
+import { getFromStorage, setInStorage } from './storage';
 
 export function getWebsites() {
     return getFromStorage('blockedUrls').then(res => res.blockedUrls || []);
@@ -91,15 +56,6 @@ export const unblockWebsite = (hostname) => {
 
         return setInStorage({ blockedUrls });
     }).then(() => message.success(`Unblocked ${hostname}`));
-};
-
-export const addStorageListener = callback => {
-    if (!(chrome && chrome.storage)) {
-        listeners.push(callback);
-        window.addEventListener('storage', callback); // only for external tab
-    } else {
-        chrome.storage.onChanged.addListener(callback);
-    }
 };
 
 export const appEnabled = callback => {
